Added integration testing
