+++ core-kotlin-modules/core-kotlin-advanced-4/src/test/kotlin/com/baeldung/runKotlinScripts/RunKotlinScriptsUnitTest.kt
@@ -8,16 +8,6 @@ import java.io.PrintStream
 import javax.script.ScriptEngineManager
 
 class RunKotlinScriptsUnitTest {
-    @Test
-    fun `run Kotlin script with ProcessBuilder`() {
-        val scriptFile = File.createTempFile("test", ".kts")
-        scriptFile.writeText("""
-        println("Hello, ProcessBuilder!")
-    """.trimIndent())
-
-        val output = runScriptUsingProcess(scriptFile.absolutePath).trim()
-        assertEquals("Hello, ProcessBuilder!", output)
-    }
 
     @Test
     fun `run Kotlin script using ScriptEngine`() {
@@ -32,13 +22,6 @@ class RunKotlinScriptsUnitTest {
     }
 
 
-}
-
-fun runScriptUsingProcess(scriptPath: String): String {
-    val processBuilder = ProcessBuilder("kotlinc", "-script", scriptPath)
-    val process = processBuilder.start()
-
-    return process.inputStream.bufferedReader().use { it.readText() }
 }
 
 fun runKotlinScriptWithEngine(scriptPath: String): String {
